+++ modules/util/trigger_event.js
@@ -1,10 +1,7 @@
 export function utilTriggerEvent(target, type, eventProperties) {
     target.each(function() {
-        var evt = document.createEvent('HTMLEvents');
-        evt.initEvent(type, true, true);
-        for (var prop in eventProperties) {
-            evt[prop] = eventProperties[prop];
-        }
+        var evt = new Event(type, { bubbles: true, cancelable: true });
+        Object.assign(evt, eventProperties);
         this.dispatchEvent(evt);
     });
 }